refactor: extract pose studio core logic, IK solvers, and Three.js module loading into a dedicated `vnccs_pose_studio_core.js` file.
